requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
tensorflow
pillow
google-generativeai
python-dotenv
